+++ galapy/analysis/__init__.py
@@ -1,0 +1,3 @@
+from galapy.analysis import model_comparison
+from galapy.analysis import plot
+from galapy.analysis import funcs